Reorganize file and folder structure for Shopping Cart components, fix routes and remove breadcrumb from shopping cart page
