+++ frontend/src/features/permits/components/cart/EditCartItemDialog.scss
@@ -1,6 +1,6 @@
-@import "../../../../../themes/orbcStyles";
+@import "../../../../themes/orbcStyles";
 
-.update-cart-dialog {
+.edit-cart-item-dialog {
   &__header {
     display: flex;
     align-items: center;
@@ -39,9 +39,25 @@
   }
 
   & &__btn {
-    color: $white;
-    font-weight: bold;
-    padding: 0.75rem 2rem;
+    &--cancel {
+      margin-right: 1.5rem;
+      color: $bc-black;
+      padding: 0.75rem 1rem;
+      background-color: $bc-background-light-grey;
+      box-shadow: none;
+
+      &:hover {
+        border: 2px solid $bc-text-box-border-grey;
+        background-color: $bc-background-light-grey;
+        cursor: pointer;
+        box-shadow: none;
+      }
+    }
+
+    &--edit {
+      color: $white;
+      font-weight: bold;
+      padding: 0.75rem 2rem;
+    }
   }
 }
-
