+++ frontend/src/features/permits/pages/ShoppingCart/components/CartChangedWarningBanner.tsx
