+++ frontend/src/features/permits/pages/ShoppingCart/components/RemoveCartButton.scss
