+++ frontend/src/features/permits/pages/ShoppingCart/components/RemoveCartButton.tsx
