+++ frontend/src/features/permits/pages/ShoppingCart/components/ShoppingCart.scss
@@ -1,4 +1,4 @@
-@import "../../../../themes/orbcStyles";
+@import "../../../../../themes/orbcStyles";
 
 .shopping-cart {
   .cart-changed-warning-banner {
