+++ frontend/src/features/permits/pages/ShoppingCart/components/ShoppingCartItem.tsx
